requirements.txt
# django
Django>=2.2,<3
django-environ==0.4.5
django-cors-headers==2.5.2

# wagtail
wagtail>=2.7,<2.8
wagtail-grapple>=0.4,<0.5
wagtailcloudinary==1.0.0

# database
psycopg2>=2.7,<3.0
dj-database-url==0.5.0

# server
gunicorn==20.0.4
whitenoise==5.0.1

# http
requests==2.22.0

# monitoring
raven>=3
